Move gaps definitions into const
